create new package Symfony, move Bridge there
