+++ packages/Symfony/src/config/services.yml
@@ -1,0 +1,6 @@
+services:
+    _defaults:
+        autowire: true
+
+    Rector\Symfony\:
+        resource: '../'
