+++ src/Rector/Contrib/Symfony/Controller/ActionSuffixRemoverRector.php
@@ -3,7 +3,7 @@
 namespace Rector\Rector\Contrib\Symfony\Controller;
 
 use PhpParser\Node;
-use Rector\Bridge\Symfony\NodeAnalyzer\ControllerMethodAnalyzer;
+use Rector\Symfony\Bridge\NodeAnalyzer\ControllerMethodAnalyzer;
 use Rector\Builder\IdentifierRenamer;
 use Rector\Rector\AbstractRector;
 use Rector\RectorDefinition\CodeSample;
